refactor(comp/process/profiler): migrate to V2 component pattern
Migrates comp/process/profiler from V1 (fx.In dependencies struct, Module() in impl) to V2 (Requires/Provides structs, def/impl/fx layout, compdef.Lifecycle).
