+++ comp/process/profiler/def/component.go
